+++ src/libs/types.h
@@ -20,9 +20,9 @@ namespace image_utils {
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wunused-variable"
     /*keep these constants here for conveniece even if they aren't used*/
-    const long double PI = 3.14159265358979323846264338327950288419716939937510582097494459230781640628620899862803482L;
-    const long double Pi = PI;
-    const long double pi = PI;
+    const double PI = 3.14159265358979323846264338327950288419716939937510582097494459230781640628620899862803482L;
+    const double Pi = PI;
+    const double pi = PI;
 #pragma clang diagnostic pop
 
     template<typename T>
